rename: 파일명 수정 gameRecord -> game-record
