+++ src/renderer/components/summoner-league-voice-block/game-record.tsx
