+++ src/renderer/components/summoner-voice-block/game-record.tsx
